requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
dj-database-url==2.1.0
Django==5.0.4
django-admin-soft-dashboard==1.0.12
django-cors-headers==4.3.1
djangorestframework==3.15.1
idna==3.7
pillow==10.3.0
psycopg2==2.9.9
python-dotenv==1.0.1
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
whitenoise==6.6.0
